refactor: abstracted app states
leader board
game instructions modal
leader board modal
