+++ src/App.tsx
@@ -2,16 +2,19 @@ import { BrowserRouter } from 'react-router-dom';
 
 import { AppRoutes } from './AppRoutes';
 import { RootLayout } from './Layouts/RootLayout';
+import { AppProvider } from './store/AppProvider';
 import { AuthProvider } from './store/AuthProvider';
 
 function App() {
     return (
         <BrowserRouter>
-            <AuthProvider>
-                <RootLayout>
-                    <AppRoutes />
-                </RootLayout>
-            </AuthProvider>
+            <AppProvider>
+                <AuthProvider>
+                    <RootLayout>
+                        <AppRoutes />
+                    </RootLayout>
+                </AuthProvider>
+            </AppProvider>
         </BrowserRouter>
     );
 }
