+++ src/Layouts/RootLayout.tsx
@@ -4,11 +4,11 @@ import OutsideClickHandler from 'react-outside-click-handler';
 
 import { Footer } from '../components/Footer/Footer';
 import { Navbar } from '../components/Navbar/Navbar';
-import { useAuthContext } from '../hooks/useAuthContext';
+import { useAppContext } from '../hooks/useAppContext';
 import LeaderBoard from '../Pages/LeaderBoard';
 
 export const RootLayout = ({ children }: { children: ReactNode }) => {
-    const { isLeaderBoardClicked, handleCloseLeaderBoard } = useAuthContext();
+    const { isLeaderBoardClicked, handleCloseLeaderBoard } = useAppContext();
 
     return (
         <section className='flex h-screen flex-col'>
