+++ src/hooks/useAuthContext.ts
@@ -3,31 +3,11 @@ import { useContext } from 'react';
 import { AuthContext } from '../store/AuthContext';
 
 export const useAuthContext = () => {
-    const {
-        user,
-        leaderBoardArr,
-        isAuthenticated,
-        isGameModalClicked,
-        isLeaderBoardClicked,
-        setLeaderBoardArr,
-        setIsAuthenticated,
-        handleOpenGameModal,
-        handleCloseGameModal,
-        handleOpenLeaderBoard,
-        handleCloseLeaderBoard,
-    } = useContext(AuthContext);
+    const { user, isAuthenticated, setIsAuthenticated } = useContext(AuthContext);
 
     return {
         user,
-        leaderBoardArr,
         isAuthenticated,
-        isGameModalClicked,
-        isLeaderBoardClicked,
-        setLeaderBoardArr,
         setIsAuthenticated,
-        handleOpenGameModal,
-        handleCloseGameModal,
-        handleOpenLeaderBoard,
-        handleCloseLeaderBoard,
     } as const;
 };
